we now produce the right output

+++ src/tip/calculatetip.go
@@ -2,7 +2,7 @@ package tip
 
 import "math"
 
-func CalculateTip(amountInCents int, tipPercentage int) (int, int) {
-	tipAmount := int(math.Ceil(float64(amountInCents) * float64(tipPercentage) / 100.0))
+func CalculateTip(amountInCents int64, tipPercentage int64) (int64, int64) {
+	tipAmount := int64(math.Ceil(float64(amountInCents) * float64(tipPercentage) / 100.0))
 	return amountInCents + tipAmount, tipAmount
 }
